+++ scim2_cli/query.py
@@ -11,6 +11,7 @@ from sphinx_click.rst_to_ansi_formatter import make_rst_to_ansi_formatter
 
 from .utils import DOC_URL
 from .utils import formatted_payload
+from .utils import split_headers
 
 
 @click.command(cls=make_rst_to_ansi_formatter(DOC_URL), name="query")
@@ -49,6 +50,9 @@ from .utils import formatted_payload
     help="A string indicating the order in which the “sortBy” parameter is applied.",
 )
 @click.option(
+    "-h", "--headers", multiple=True, help="Header to pass in the HTTP requests."
+)
+@click.option(
     "--indent/--no-indent",
     is_flag=True,
     default=True,
@@ -65,6 +69,7 @@ def query_cli(
     filter: str,
     sort_by: str,
     sort_order: str,
+    headers: List[str],
     indent: bool,
 ):
     """Perform a `SCIM GET <https://www.rfc-editor.org/rfc/rfc7644#section-3.4.1>`_ request
@@ -113,11 +118,14 @@ def query_cli(
                 id,
                 search_request=payload,
                 check_request_payload=check_request_payload,
+                headers=split_headers(headers),
             )
 
         else:
             response = ctx.obj["client"].query_all(
-                search_request=payload, check_request_payload=check_request_payload
+                search_request=payload,
+                check_request_payload=check_request_payload,
+                headers=split_headers(headers),
             )
 
     except (httpx.HTTPError, SCIMClientError) as exc: