+++ doc/changelog.rst
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+Added
+^^^^^
+- Add support for passing custom headers to requests.
+
 [0.1.1] - 2024-06-02
 --------------------